better structured code for the strip lights
